Added a bunch of stuff. #4 #5
Added modmenu and clothconfig2 integration, as well as a almost decent settings menu _and_ settings files.
 - config.properties will store all straight forward settings like integers strings and alike.
 - enchantment_data.json currently stores all the information related to enchantment colors and index values.

The EnchantmentData structure was heavily changed during the previously mentioned additions and it may change even more in the future since this was never set in stone and organic evolution for such data classes is very beneficial.

With this tho unfortunately the sorting broke. Not sure when or why but it sure did, probably a lot will have to be changed to fix this.