Big Port
This is a chonker
Lots of changes, I tried to go back and split this into bite-sized commits, but stuff was too intertwined to make it worthwhile.
Most notably, I've prepared a new way to save and load individual enchantment settings to/from disk using kotlinx.serialization. This is not implemented yet because I still want to work out backwards compatibility (reading from the old format).
I also managed to bring most utility methods to kotlin as well and converted them to extension methods since I do like those more as a personal preference.
The original modconfig was removed as I moved to the new EnchantmentData format.